Extract to Highlighter component
