+++ src/renderer/components/NodeDocumentation/docsMarkdown.tsx
@@ -1,39 +1,13 @@
-import { Code, Highlight, Link, Text } from '@chakra-ui/react';
+import { Code, Link, Text } from '@chakra-ui/react';
 import ChakraUIRenderer from 'chakra-ui-markdown-renderer';
-import { Fragment, PropsWithChildren, memo } from 'react';
+import { PropsWithChildren, memo } from 'react';
 import { Components } from 'react-markdown';
 import { useContext } from 'use-context-selector';
 import { SchemaId } from '../../../common/common-types';
 import { BackendContext } from '../../contexts/BackendContext';
 import { NodeDocumentationContext } from '../../contexts/NodeDocumentationContext';
+import { Highlighter } from './Highlighter';
 import { SchemaLink } from './SchemaLink';
-
-const recursivelyMarkChildren = (
-    children: React.ReactNode | string[],
-    searchTerms: readonly string[]
-): React.ReactNode | string[] => {
-    if (Array.isArray(children)) {
-        return children.map((child, i) => (
-            // eslint-disable-next-line react/no-array-index-key
-            <Fragment key={i}>{recursivelyMarkChildren(child, searchTerms)}</Fragment>
-        ));
-    }
-
-    if (typeof children === 'string') {
-        return (
-            <Highlight
-                query={[...searchTerms]}
-                styles={{
-                    backgroundColor: 'yellow.300',
-                }}
-            >
-                {children}
-            </Highlight>
-        );
-    }
-
-    return children;
-};
 
 const getDocsMarkdownComponents = (interactive: boolean): Components => {
     return {
@@ -47,7 +21,7 @@ const getDocsMarkdownComponents = (interactive: boolean): Components => {
                     marginTop={1}
                     userSelect="text"
                 >
-                    {recursivelyMarkChildren(children, searchTerms)}
+                    <Highlighter searchTerms={searchTerms}>{children}</Highlighter>
                 </Text>
             );
         }),
